Merged in python_wrapper2 (pull request #3)
Python wrapper2

* wrapper_utils.py thoroughly tested. No error handling done, but that's a wish for later. I'm writing the docstrings.

* First complete version of docstrings finished

* wrapper_utils.py with proper typing done and docstrings that reflect it.

read_input_files() in wrapper_utils.py accepts list of files for a fully pythonic experience.

calculator.py now checks the length of the protein so the progressbar can show a percentage, and I changed the unit to "residues" instead of iterations

methods.py now has a default random seed for the boostrap to ensure reproducibility in the sampling, probably quite important for low bootstraps.

Finally, data/__init__.py was added to be able to store KDEs in a default directory.

* Good beginning of README.md but it definitelly needs another look.

Bug with calculate_results in wrapper_utils.py is fixed and now it can calculate all windows and bootstraps as needed.

* Some changes suggested on the pull request are implemented. Waiting for reply on some of my replies and I'll fix the rest.

* Changes from the pull requests are implemented. I had to include the data directory inside constava because otherwise data was stored at the same level as any other package. Some other files were modified to reflect this change.

Now the seed for bootstrap is by default None and it can be changed from the wrapper.

* Further changes from the pull request implemented

* Revert changes to methods.py

* Delete redundant data-dir